Fixed not gate bug
The bug was that some not gates (I think if the input wires were picked
in a particular way) would output a random value - sometimes 0 and
sometimes 1 - independently of the inputted value. This fix forces the
not gate to only consider a single wire, as opposed to the intended
input wire and wire 0.